Intel RDT: updates according to proposed spec changes.
There are two proposed clarifications to the OCI spec.

The first item specifies that the L3CacheSchema line filtering needs to
happen only when both MemBw and L3Cache schemas are specified.

The second item specifies that the subdirectory needs to be deleted.
Runc already does that, but the clarification adds for directory removal
only if the directory was created by us.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>